Stage 1 of 3 for JetBrains Academy - Kotlin Android - Tip Calculator project. Objectives Add the following components to the application layout: 1. EditText field where the user enters the bill value with the id attribute edit_text and the inputType attribute number; 2. Slider widget for specifying the tip percentage with the id attribute slider; the values range from 0 to 100 with the step size 1; 3. An empty TextView to later display the calculation result with the id attribute text_view.
This is my first time using Git for Android Studios, so I'm not exactly sure I'm saving all the necessary files ( there seems to be a ridiculous amount of files, so I just picked the ones that seemed most relevant. Like for gradle, I think all that is needed are the two build.gradle files, but I'm not 100% sure, and some files seemed specific to my computer so I thought it best to leave those out ).

activity_main.xml changes:
• I changed the default ConstraintLayout to a vertical LinearLayout to help conform to the layout they wanted.
• added an EditText limited to just numbers for input and id edit_text
• added a TextView with id text_view
• added a Slider with id slider, step size of 1, and range from 0 to 100.