Stage 2 of 3 for JetBrains Academy - Kotlin Android - Tip Calculator project. Description
Now that your application has the appropriate UI components, it needs to be able to tell when the user input changes: this way, it will know when to redo the calculation and show the updated tip amount. For now, we will use the EditText field to show the user input values.
Objectives

Add the appropriate event listeners to the EditText field and the Slider so that the program knows when the user provides input.

Show the user input values in the TextView in the format Bill value: <X>, tip percentage: <P>% with the corresponding values instead of <X> and <P>.

Using the event listeners, update TextView each time the user input changes. If the bill value is not provided (that is, EditText is empty), then TextView should disappear.

MainActivity changes:
• added val calculator that links to CalculatorViewModel
• observe calculator.getResult to update the view named text_view
• set doOnTextChanged for view edit_text, which calls calculator.billChange() with the updated data.
• set addOnChangeListener for view slider, which calls calculator.percentChange() with the updated data.