Stage 3 of 3 for JetBrains Academy - Kotlin Android - Tip Calculator project.
I found on stackoverflow the files that are needed for here. I also had to move files around, as it was a class project, so the files weren't where they needed to be to build it as an app ( next class project I'll have two projects, one for uploading to here, and one for the class ).